Add "ansible_host" to each host
Changes to the lxc_container_create role caused the swift tests to begin
failing. The "Wait for ssh to be available" task requires an
"ansible_host" value, which was not set for the Swift test containers,
as a result this task continuously failed despite the containers being
up and accessible.

Change-Id: I69e8f9bbc0c0a69a24b4a4143c806ea10779f0b0