Swift proxy performance fixes
Performance testing of an OSA deployed Swift cluster hit several issues.
Comparison of the proxy conf with the Swift Auth doc [1] found
two settings that affect performance that that were not set in the template.

include_service_catalog - This defaults to True which makes the proxy server
fetch the service catalog on every request in the authtoken middleware. This
hammers Keystone.

cache - When using memcache this should be set so the authtoken middleware
uses the correct cache and not go back to Keystone on every request.

[1] http://docs.openstack.org/developer/swift/overview_auth.html

Change-Id: Iddaf097e3ff5c1b7118b11cfc598a40a91d6f96f
(cherry picked from commit b0f330e7b2a805c87a988d1cd312ec766e5e9f39)