Add object-reconstructor for swift EC
Swift Erasure Coding requires the object-reconstructor to be running,
this PR adds the object-reconstructor configuration and systemd/upstart
file.

Additionally, this PR adds testing for an erasure-coding policy as well
as for multiple policies in 1 deployment.

This is a manual re-implementation of [1] for the Mitaka branch due to
the differences in the service contiguration code between Ocata/Newton
and Mitaka.

[1] https://review.openstack.org/#/c/388819

Change-Id: I37934da8483d997dd669e0317be09596fed657e6
Closes-Bug: #1634859