Skip with_first_found when no vars file is found
The upgrade tests we run in Ocata will run against stable/newton where
the Redhat7 vars file does not exist, since we don't have Redhat7
support.

Due to the way includes happen, the files are still evaluated, which
means the task will fail. To avoid this we don't fail on the
"with_first_found" task when no file exists.

After stable/newton the Redhat7 vars will be found, so we don't need
this change in Ocata+.

Change-Id: I5d30fdcdc901e13623efff9901acb20ab24c3bab