Move swift upgrades into its own shell script
The upgrade path becomes complicated as previous variables are stored
and we have to start a clean run.

The best way to achieve this is to separate the run steps out in a shell
script for upgrades only.

Change-Id: Ib2d2c8bad5241779d52cdc9428d7057c770d2d7c
(cherry picked from commit 44cdfed99de629d6ba6df0d5abbc83ead7d7fd9e)