Fix ordering of swift3 in middleware pipeline
Beginning with commit 6ffcc29 of swift3, the swift3 and s3token middlewares
must come between authtoken and keystoneauth in the swift proxy pipeline.

When 6ffcc29 was committed, reordering was optional at the expense of an
unneceessary call to keystone.  However, the change becomes mandatory when using
keystone v3 tokens.  Without this change, authtoken will strip the necessary
headers that s3token just added to the environment.

Change-Id: Ic7d07d869aa617ee00190a9aedf411017539b97b