Optimize swift gate
This patch contains five combined improvements.

1) Remove RabbitMQ from the gate

RabbitMQ is only needed in the swift gate if the ceilometer
middleware is being deloyed. This isn't enabled in the gate
job, so let's save some time and deploy memcached/galera
without RabbitMQ.

2) Bindep pyOpenSSL change

This commit also includes the pyOpenSSL bindep change in the
hopes that we can get the swift gate moving along again. ;)

3) Disable erasure code test for CentOS

Due to #1707220 we are running into issues with CentOS 7 running Erasure
Code storage policies on RAX hardware.

This patch temoprarily disables EC testing for CentOS 7, but we can
re-enable that once the package is backported to stable/ocata branch.

4) Pass packages as a list

This patch passes a list of packages to the package module
rather than using "with_items" and it should speed up the
package installation process.

5) Set package state to present

Change-Id: Ic7b13853ce00b0f8ac2181e46e2bd8dea5ce377e
Co-Authored-By: Andy McCrae <andy.mccrae@gmail.com>