Initial OSA zuul v3 role jobs
This patch implements an initial set of jobs intended to match
the current job execution method. It does not intend to improve
how the jobs are executed - only to replicate what is currently
in openstack-infra/openstack-zuul-jobs and provide the platform
to iterate on.

Co-Authored-By: Jean-Philippe Evrard <jean-philippe@evrard.me>
Change-Id: I4aa7591c931501490e706b9d61ba0bc12c006baa
(cherry picked from commit 0147d155d467f8dfbe3820e92cd9bd657b8d4974)