Only replace python when re-initializing the venv
We reinitialize the venv to ensure that the right version of
python is in the venv, but we do not want virtualenv to also
replace pip, setuptools and wheel so we tell it not to. If we
don't do this then virtualenv will install the latest available
version, which is not what we want.

Change-Id: I05d9ee40e596afcc35559b53f6a773521b1d1534
Partial-Bug: #1764470
(cherry picked from commit 1a789263bd15a44a4e9774ee5dd57b634e4e3676)