Enable overriding the service setup host python interpreter
In order to enable the service setup host python interpreter to
be changed easily, we make it a variable. This will be useful
when someone sets the service setup host to be the utility
container, because we'll be able to set this var by default.

Change-Id: Icb3ea9ad782218c357ec0e3c577ecbe6fbf60461
(cherry picked from commit c7e11dfcbda5b2984648a649403879e1b3e59c99)